backend_temp/requirements.txt
Django>=4.2
mysqlclient
python-dotenv
djangorestframework>=3.14
djangorestframework-simplejwt
django-filter
django-ckeditor
Pillow 